mps_cli/requirements.txt
supabase==2.3.4
python-dotenv==1.0.0
wcwidth==0.2.12